refactor(engrave): drop heuristic fallback, always hit ML service
The local music21 engrave path produced poor output; keeping it as a
silent fallback only masked problems with the ML service. With the
seq2seq engraver now the single source of truth, the toggle becomes
operational dead weight.

- Delete backend/services/engrave.py + backend/workers/engrave.py.
  Extract the one piece the runner still needs (render_midi_bytes)
  into backend/services/midi_render.py.
- Remove OHSHEET_ENGRAVER_INFERENCE and the use_ml_engraver branching
  in PipelineRunner — audio_upload / midi_upload jobs unconditionally
  hit the ML service via ml_engraver_client.
- Hard-fail title_lookup jobs that reach the engrave stage: those are
  expected to resolve upstream via TuneChat; with no local fallback,
  silent failure is worse than a clear error.
- Drop worker-engrave from docker-compose{,-prod}.yml; it has nothing
  to run.
- /v1/artifacts/{job}/pdf returns 404 now — PDF rendering is a client
  responsibility, the ML engraver returns MusicXML only.
- Update tests: stub the ml engraver in conftest so pipeline tests
  don't need a live backend; delete the quality / pdf-render / toggle
  suites that exercised music21 paths.
