+++ .github/workflows/deploy.yml
@@ -96,10 +96,9 @@ jobs:
           OHSHEET_TUNECHAT_API_KEY: ${{ secrets.OHSHEET_TUNECHAT_API_KEY }}
           # Anthropic API — used by worker-refine for LLM metadata enrichment.
           OHSHEET_ANTHROPIC_API_KEY: ${{ secrets.OHSHEET_ANTHROPIC_API_KEY }}
-          # Engraver inference toggle — replaces the retired OHSHEET_ML_API_URL gate.
-          # OHSHEET_ENGRAVER_INFERENCE: bool gate read by backend.jobs.runner.
-          # OHSHEET_ENGRAVER_SERVICE_URL: base URL of oh-sheet-ml-pipeline engraver service.
-          OHSHEET_ENGRAVER_INFERENCE: ${{ vars.OHSHEET_ENGRAVER_INFERENCE }}
+          # Base URL of oh-sheet-ml-pipeline engraver service. The engrave
+          # stage always hits this URL — there is no local fallback — so
+          # operators rely on service health, not a toggle.
           OHSHEET_ENGRAVER_SERVICE_URL: ${{ vars.OHSHEET_ENGRAVER_SERVICE_URL }}
         run: |
           # Set up SSH
@@ -124,7 +123,6 @@ jobs:
           echo "OHSHEET_TUNECHAT_API_KEY=${OHSHEET_TUNECHAT_API_KEY}" >> /tmp/.env
           # Anthropic — worker-refine LLM metadata enrichment.
           echo "OHSHEET_ANTHROPIC_API_KEY=${OHSHEET_ANTHROPIC_API_KEY}" >> /tmp/.env
-          echo "OHSHEET_ENGRAVER_INFERENCE=${OHSHEET_ENGRAVER_INFERENCE}" >> /tmp/.env
           echo "OHSHEET_ENGRAVER_SERVICE_URL=${OHSHEET_ENGRAVER_SERVICE_URL}" >> /tmp/.env
 
           # Copy compose files to VM (SCP .env to temp name, then atomic mv)
