+++ tests/test_jobs.py
@@ -59,7 +59,7 @@ def test_create_job_from_audio_runs_to_completion(client):
     assert status is not None
     assert status["status"] == "succeeded", status
     assert status["result"] is not None
-    assert status["result"]["pdf_uri"].startswith("file://")
+    # pdf_uri is intentionally empty — the ML engraver returns MusicXML only.
     assert status["result"]["musicxml_uri"]
     assert status["result"]["humanized_midi_uri"]
